Bump iota.rs commit (#1013)
* Bump iota.rs commit

* Set allow_burning to false

* Update Cargo.toml

+++ src/account/operations/helpers/time.rs
@@ -20,11 +20,11 @@ use crate::{
 impl AccountHandle {
     /// Get the local time, but compare it to the time from the nodeinfo, if it's off more than 5 minutes, an error will
     /// be returned
-    pub(crate) async fn get_time_and_milestone_checked(&self) -> Result<(u64, u32)> {
+    pub(crate) async fn get_time_and_milestone_checked(&self) -> Result<(u32, u32)> {
         let local_time = SystemTime::now()
             .duration_since(UNIX_EPOCH)
             .expect("Time went backwards")
-            .as_secs();
+            .as_secs() as u32;
         let status_response = self.client.get_info().await?.nodeinfo.status;
         let latest_ms_timestamp = status_response.latest_milestone_timestamp;
         // Check the local time is in the range of +-5 minutes of the node to prevent locking funds by accident
